chg: test: move tumblr method test !refactor
